feat(samples): Updating and adding samples for pnp summer refresh (#841)
