+++ service/iot-service-samples/pnp-service-sample/pom.xml
@@ -10,6 +10,18 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>pnp-service-sample</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>thermostat-service-sample</module>
+        <module>temperature-controller-service-sample</module>
+    </modules>
 
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <lombok-version>1.18.8</lombok-version>
+        <gson.version>2.8.6</gson.version>
+        <slf4j-version>1.7.25</slf4j-version>
+        <maven-jar-plugin-version>3.1.0</maven-jar-plugin-version>
+    </properties>
 
 </project>
